set up file infrastructure, working on budget input and view
