+++ src/main/java/com/budget/budgettracking/BudgetView.java
@@ -1,0 +1,9 @@
+package com.budget.budgettracking;
+
+import javafx.scene.control.Tab;
+
+public class BudgetView extends Tab {
+    // tab for BudgetInput tabpane
+    // user can view visual representation of budget piechart and metrics
+    // returns tab
+}
