+++ src/main/java/com/budget/budgettracking/DataStorage.java
@@ -1,0 +1,4 @@
+package com.budget.budgettracking;
+
+public class DataStorage {
+}
