+++ src/main/java/com/budget/budgettracking/Main.java
@@ -1,0 +1,4 @@
+package com.budget.budgettracking;
+
+public class Main {
+}
