+++ src/main/java/com/budget/budgettracking/TransactionInput.java
@@ -1,0 +1,6 @@
+package com.budget.budgettracking;
+
+public class TransactionInput {
+    // window where user can input transactions and see table
+    // returns a scene
+}
