+++ src/main/java/com/budget/budgettracking/TransactionView.java
@@ -1,0 +1,10 @@
+package com.budget.budgettracking;
+
+import javafx.scene.control.Tab;
+
+public class TransactionView extends Tab {
+    // tab for TransactionInput tabpane
+    // user can view visualization of transactions, pie/chart/line, list/table by time
+    // returns tab
+
+}
