+++ src/main/resources/com/budget/budgettracking/stylesheet.css
@@ -1,3 +1,8 @@
+@font-face {
+    font-family: 'Roboto-Regular';
+    src: url('fonts/Roboto-Regular.ttf')
+}
+
 .chart-pie-label-line {
     -fx-stroke: black;
     -fx-fill: black;
@@ -10,4 +15,22 @@
 .chart-legend {
    -fx-background-color:  #fafad2;
    -fx-stroke: #ce9c1d;
+}
+
+.custom-label {
+    -fx-font-family: Roboto-Regular;
+    -fx-font-size: 14px;
+    -fx-font-weight: bold;
+}
+
+.button {
+    -fx-font-family: Roboto-Regular;
+    -fx-font-size: 14px;
+    -fx-font-weight: bold;
+}
+
+.title {
+    -fx-font-family: Roboto-Regular;
+    -fx-font-size: 18px;
+    -fx-font-weight: bold;
 }
